+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - overlay-manager-rc@0.7.1
+
 ## 1.0.3
 
 ### Patch Changes